style: normalize pre-existing rustfmt drift and fix latent clippy lints
The committed tree did not pass `cargo fmt --check` (formatting was never
a gated check — SPEC.md lists test/clippy/rustdoc but not fmt), and three
files carried clippy violations under rustc 1.94.1:

- field_reassign_with_default in the 問257 non-manifold tests
  (extract/mesh.rs, verify/check.rs) — rewritten as struct literals.
- an unnecessary u8->u8 cast in the DEFLATE stored-block decoder
  (render/deflate.rs).

All changes are logic-preserving. Kept separate from the chamfer feature
commit so each is coherent.

+++ src/render/deflate.rs
@@ -403,7 +403,7 @@ mod tests {
                     let len = br.read_bits(16) as usize;
                     let _nlen = br.read_bits(16);
                     for _ in 0..len {
-                        out.push(br.data[br.byte_pos] as u8);
+                        out.push(br.data[br.byte_pos]);
                         br.byte_pos += 1;
                     }
                 }
